requirements.txt
pip>=24.0
setuptools>=70.0
wheel>=0.42
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-WTF==1.1.1
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
qrcode==7.4.2
Pillow>=10.2.0
Werkzeug==2.3.7
WTForms==3.0.1
python-dateutil==2.8.2
google-api-python-client==2.100.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.1.0
email-validator==2.1.1
boto3
XlsxWriter>=3.1.0
sendgrid
scikit-learn
pandas
openpyxl
requests==2.31.0
postmark
postmarker
celery
redis
Flask-Caching
